+++ src/pages/Lend/LoanChannelCategoryControl.vue
@@ -62,6 +62,7 @@
 							:show-action-button="true"
 							:show-tags="enableLoanTags"
 							:in-grid="true"
+							:category-page-name="loanChannelName"
 						/>
 						<loan-card-controller
 							v-else
@@ -96,6 +97,7 @@
 							:show-action-button="true"
 							:show-tags="enableLoanTags"
 							:in-grid="true"
+							:category-page-name="loanChannelName"
 						/>
 						<loan-card-controller
 							v-else
@@ -130,6 +132,7 @@
 							:show-action-button="true"
 							:show-tags="enableLoanTags"
 							:in-grid="true"
+							:category-page-name="loanChannelName"
 						/>
 						<loan-card-controller
 							v-else